Make functions for executing run command reusable
