+++ pkg/component/validate.go
